#36 fix: 사용자 중복투표 방지 기능 구현 + 전체 투표수 출력

+++ server/src/main/java/com/seb40/server/Answer/Entity/Answer.java
@@ -26,7 +26,7 @@ public class Answer {
     @JoinColumn(name ="user_id")
     private User user;
 
-    private int answerVoteSum;
+    private long answerVoteSum;
 
     public void setUser(User user){this.user = user;}
